3DS port of FrodoDS.
You need a way to run homebrew applications on your 3DS. See
http://smealum.github.io/3ds/ for details.

Be sure to place BreadBox.3dsx and BreadBox.smdh in /3ds/BreadBox/
Any games you with to load should be placed in /c64/games/

Thats it. Currently when selecting a game, the emulator will try to
load it
using Load"*",8,1 however, this does not always work.